refactor and test lsr service to always emit data
https://mesonet.agron.iastate.edu/cgi-bin/request/gis/lsr.py?help
